Added: tests partial update user location
